Scrap build.js too
