+++ src/base65537.js
@@ -2,9 +2,7 @@
   Oh boy is this slow.
 */
 
-const build = require('./build.js')
-
-const chrs = [
+const repertoire = [
   '㐀㔀㘀㜀㠀㤀㨀㬀㰀㴀㸀㼀䀀䄀䈀䌀䐀䔀䘀䜀䠀䤀䨀䬀䰀一伀倀儀刀匀吀唀嘀圀堀夀娀嬀尀崀帀开怀愀戀挀搀攀昀最栀椀樀欀氀洀渀漀瀀焀爀猀琀甀瘀眀砀礀稀笀簀紀縀缀耀脀舀茀萀蔀蘀蜀蠀褀言謀谀贀踀輀退鄀鈀錀鐀销阀需頀餀騀鬀鰀鴀鸀ꄀꈀꌀꔀ𐘀𒀀𒄀𒈀𓀀𓄀𓈀𓌀𔐀𔔀𖠀𖤀𠀀𠄀𠈀𠌀𠐀𠔀𠘀𠜀𠠀𠤀𠨀𠬀𠰀𠴀𠸀𠼀𡀀𡄀𡈀𡌀𡐀𡔀𡘀𡜀𡠀𡤀𡨀𡬀𡰀𡴀𡸀𡼀𢀀𢄀𢈀𢌀𢐀𢔀𢘀𢜀𢠀𢤀𢨀𢬀𢰀𢴀𢸀𢼀𣀀𣄀𣈀𣌀𣐀𣔀𣘀𣜀𣠀𣤀𣨀𣬀𣰀𣴀𣸀𣼀𤀀𤄀𤈀𤌀𤐀𤔀𤘀𤜀𤠀𤤀𤨀𤬀𤰀𤴀𤸀𤼀𥀀𥄀𥈀𥌀𥐀𥔀𥘀𥜀𥠀𥤀𥨀𥬀𥰀𥴀𥸀𥼀𦀀𦄀𦈀𦌀𦐀𦔀𦘀𦜀𦠀𦤀𦨀𦬀𦰀𦴀𦸀𦼀𧀀𧄀𧈀𧌀𧐀𧔀𧘀𧜀𧠀𧤀𧨀𧬀𧰀𧴀𧸀𧼀𨀀𨄀𨈀𨌀𨐀𨔀',
   '㐁㔁㘁㜁㠁㤁㨁㬁㰁㴁㸁㼁䀁䄁䈁䌁䐁䔁䘁䜁䠁䤁䨁䬁䰁丁企倁儁刁匁吁唁嘁圁堁夁威嬁封崁币弁态愁戁持搁攁昁朁栁椁樁欁氁洁渁漁瀁焁爁猁琁甁瘁省码礁稁笁簁紁縁缁老脁舁茁萁蔁蘁蜁蠁褁訁謁谁贁踁輁送鄁鈁錁鐁锁阁霁頁餁騁鬁鰁鴁鸁ꄁꈁꌁꔁ𐘁𒀁𒄁𒈁𓀁𓄁𓈁𓌁𔐁𔔁𖠁𖤁𠀁𠄁𠈁𠌁𠐁𠔁𠘁𠜁𠠁𠤁𠨁𠬁𠰁𠴁𠸁𠼁𡀁𡄁𡈁𡌁𡐁𡔁𡘁𡜁𡠁𡤁𡨁𡬁𡰁𡴁𡸁𡼁𢀁𢄁𢈁𢌁𢐁𢔁𢘁𢜁𢠁𢤁𢨁𢬁𢰁𢴁𢸁𢼁𣀁𣄁𣈁𣌁𣐁𣔁𣘁𣜁𣠁𣤁𣨁𣬁𣰁𣴁𣸁𣼁𤀁𤄁𤈁𤌁𤐁𤔁𤘁𤜁𤠁𤤁𤨁𤬁𤰁𤴁𤸁𤼁𥀁𥄁𥈁𥌁𥐁𥔁𥘁𥜁𥠁𥤁𥨁𥬁𥰁𥴁𥸁𥼁𦀁𦄁𦈁𦌁𦐁𦔁𦘁𦜁𦠁𦤁𦨁𦬁𦰁𦴁𦸁𦼁𧀁𧄁𧈁𧌁𧐁𧔁𧘁𧜁𧠁𧤁𧨁𧬁𧰁𧴁𧸁𧼁𨀁𨄁𨈁𨌁𨐁𨔁',
   '㐂㔂㘂㜂㠂㤂㨂㬂㰂㴂㸂㼂䀂䄂䈂䌂䐂䔂䘂䜂䠂䤂䨂䬂䰂丂伂倂儂刂匂吂唂嘂圂堂夂娂嬂専崂市异怂愂戂挂搂攂昂朂栂椂樂欂氂洂渂漂瀂焂爂猂琂甂瘂眂砂礂稂笂簂紂縂缂耂脂舂茂萂蔂蘂蜂蠂褂訂謂谂贂踂輂适鄂鈂錂鐂锂阂霂頂餂騂鬂鰂鴂鸂ꄂꈂꌂꔂ𐘂𒀂𒄂𒈂𓀂𓄂𓈂𓌂𔐂𔔂𖠂𖤂𠀂𠄂𠈂𠌂𠐂𠔂𠘂𠜂𠠂𠤂𠨂𠬂𠰂𠴂𠸂𠼂𡀂𡄂𡈂𡌂𡐂𡔂𡘂𡜂𡠂𡤂𡨂𡬂𡰂𡴂𡸂𡼂𢀂𢄂𢈂𢌂𢐂𢔂𢘂𢜂𢠂𢤂𢨂𢬂𢰂𢴂𢸂𢼂𣀂𣄂𣈂𣌂𣐂𣔂𣘂𣜂𣠂𣤂𣨂𣬂𣰂𣴂𣸂𣼂𤀂𤄂𤈂𤌂𤐂𤔂𤘂𤜂𤠂𤤂𤨂𤬂𤰂𤴂𤸂𤼂𥀂𥄂𥈂𥌂𥐂𥔂𥘂𥜂𥠂𥤂𥨂𥬂𥰂𥴂𥸂𥼂𦀂𦄂𦈂𦌂𦐂𦔂𦘂𦜂𦠂𦤂𦨂𦬂𦰂𦴂𦸂𦼂𧀂𧄂𧈂𧌂𧐂𧔂𧘂𧜂𧠂𧤂𧨂𧬂𧰂𧴂𧸂𧼂𨀂𨄂𨈂𨌂𨐂𨔂',
@@ -264,5 +262,55 @@ const chrs = [
   '💩'
 ].join('')
 
-module.exports = build(chrs)
-module.exports.chrs = chrs
+const chrs = [...repertoire]
+const unchrs = Object.fromEntries(Object.entries(chrs).map(([key, value]) => [value, key]))
+
+const repRange = BigInt(chrs.length)
+const byteRange = BigInt(Math.pow(2, 8))
+
+// `digits` should be an array of integers from [0, base)
+// `base` should be a BigInt
+const digitsToBigInt = (digits, base) =>
+  digits.reduce((acc, digit) => acc * base + BigInt(digit) + 1n, 0n)
+
+// `base` should be a BigInt
+const bigIntToDigits = (bigInt, base) => {
+  const digits = []
+  while (bigInt > 0n) {
+    bigInt -= 1n
+    digits.push(Number(bigInt % base))
+    bigInt /= base // rounds down
+  }
+  digits.reverse()
+  return digits
+}
+
+module.exports.encode = buf => bigIntToDigits(
+  digitsToBigInt(
+    [...buf.values()],
+    byteRange
+  ),
+  repRange
+)
+  .map(digit => {
+    if (!(digit in chrs)) {
+      throw Error(`Can't convert digit ${digit} to chr`)
+    }
+    return chrs[digit]
+  })
+  .join('')
+
+module.exports.decode = str => Buffer.from(
+  bigIntToDigits(
+    digitsToBigInt(
+      [...str].map(chr => {
+        if (!(chr in unchrs)) {
+          throw Error(`Can't convert character ${chr} to digit`)
+        }
+        return unchrs[chr]
+      }),
+      repRange
+    ),
+    byteRange
+  )
+)
