+++ test/index.js
@@ -1,2 +1,1 @@
-require('./build.js')
 require('./base65537.js')
